# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Mpango ERP API
  description: |
    Multi-tenant ERP system for African wholesale-retail operations.

    ## Authentication
    All endpoints except `/auth/login` require Bearer token authentication.
    The JWT token contains `tenant_id`, `tenant_schema`, and `user_id` claims.

    ## Multi-Tenancy
    - Login requires `tenant_code` to identify the wholesaler
    - All subsequent requests use tenant info from JWT claims
    - Data is isolated per tenant schema

    ## Authorization
    Endpoints are protected by RBAC permissions as defined in RBAC Matrix (MVP).
  version: 1.0.0
  contact:
    name: Mpango ERP Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mpango.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /users/{user_id}/roles:
    put:
      tags: [users]
      operationId: assignUserRoles
      summary: Assign roles to user
      description: Requires `roles:assign` permission
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRolesRequest'
      responses:
        '200':
          description: Roles assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # ROLE ENDPOINTS
  # ============================================================
  /roles:
    get:
      tags: [roles]
      operationId: listRoles
      summary: List all roles
      description: Requires `roles:read` permission
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================================
  # ORDER ENDPOINTS
  # ============================================================
  /orders:
    get:
      tags: [orders]
      operationId: listOrders
      summary: List orders with pagination
      description: Requires Bearer token authentication
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: retailer_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [orders]
      operationId: createOrder
      summary: Create a new order
      description: |
        Creates order with status=draft.
        `wholesaler_id` is inferred from the authenticated token (tenant_id claim).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Retailer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{order_id}:
    get:
      tags: [orders]
      operationId: getOrder
      summary: Get order by ID
      description: Requires Bearer token authentication
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{order_id}/confirm:
    post:
      tags: [orders]
      operationId: confirmOrder
      summary: Confirm order
      description: |
        Transitions order from draft → confirmed.
        Request body should be empty or `{}`.
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Order confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # PAYMENTS ENDPOINTS
  # ============================================================
...
  /orders/{order_id}/cancel:
    post:
      tags: [orders]
      operationId: cancelOrder
      summary: Cancel order
      description: |
        Can cancel from draft or confirmed status.
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderActionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


# ============================================================
# COMPONENTS
# ============================================================
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing claims:
        - user_id: UUID
        - tenant_id: UUID
        - tenant_schema: string (e.g., t_<uuid>)
        - exp: expiration timestamp

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    SizeParam:
      name: size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    UserIdParam:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OrderIdParam:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ----------------------------------------------------------
  # Reusable Responses
  # ----------------------------------------------------------
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    Pagination:
      type: object
      required: [page, size, total, pages]
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 10
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
        meta:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Input validation failed
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time

    # ========== Payment ==========
...
    User:
      type: object
      required: [id, email, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        is_active:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    AssignRolesRequest:
      type: object
      required: [role_ids]
      properties:
        role_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1

    # ========== Role ==========
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: admin
        description:
          type: string
    RoleListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        timestamp:
          type: string
          format: date-time

    # ========== Order ==========
    OrderStatus:
      type: string
      enum: [draft, confirmed, cancelled]
    OrderItemCreate:
      type: object
      required: [product_name, sku_code, quantity, unit_price]
      properties:
        product_name:
          type: string
        sku_code:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price:
          type: number
          format: float
          minimum: 0
    OrderCreateRequest:
      type: object
      required: [retailer_id, items]
      properties:
        retailer_id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemCreate'
          minItems: 1
        notes:
          type: string
    OrderItem:
      type: object
      required: [id, product_name, sku_code, quantity, unit_price, subtotal]
      properties:
        id:
          type: string
          format: uuid
        product_name:
          type: string
        sku_code:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
          format: float
        subtotal:
          type: number
          format: float
    Order:
      type: object
      required: [id, wholesaler_id, retailer_id, status, total_amount, items, created_at]
      properties:
        id:
          type: string
          format: uuid
        wholesaler_id:
          type: string
          format: uuid
        retailer_id:
          type: string
          format: uuid
        retailer_name:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        total_amount:
          type: number
          format: float
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        notes:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Order'
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    OrderListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [items, pagination]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Order'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    OrderActionResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [order_id, status]
          properties:
            order_id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/OrderStatus'
        message:
          type: string
          example: Order confirmed
        timestamp:
          type: string
          format: date-time